test: update integration test to log events and simplify flow setup
Refactors the existing real flow execution test to include minimal logging of events,
removes an obsolete test file, and streamlines the test flow creation and execution process.
Enhances visibility into flow and task lifecycle during testing.
